requirements.txt
discord.py
requests
loguru
python-dotenv